person_queue_count/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
authlib==1.6.0
avro==1.12.0
blinker==1.9.0
cachetools==5.5.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.2.1
cloudpickle==3.1.1
confluent-kafka==2.9.0
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
distro==1.9.0
docstring-parser==0.16
fastapi==0.116.0
fastavro==1.11.1
filelock==3.13.1
flask==3.1.1
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
google-adk==1.5.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.102.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.24.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.4
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.10.0
jsonlines==4.0.0
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.15
kiwisolver==1.4.8
lap==0.5.12
litellm==1.74.0.post1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mcp==1.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.3
numpy==2.1.2
openai==1.93.2
opencv-python==4.12.0.88
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
openvino==2025.2.0
openvino-telemetry==2025.2.0
orjson==3.10.18
packaging==24.1
pandas==2.3.1
pillow==11.0.0
ply==3.11
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.2
pyparsing==3.2.3
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
quixstreams==3.19.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rocksdict==0.3.27
rpds-py==0.26.0
rsa==4.9.1
scipy==1.16.0
seaborn==0.13.2
setuptools==70.2.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.4.1
starlette==0.46.2
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.66.5
typing-extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.59
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==1.26.13
uvicorn==0.35.0
websockets==15.0.1
werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
